soc: bus: Bring down reserved chunk to 256 bytes
Previously, there was 4096 bytes reserved after the IRQ table for
allowing a higher number of interrupt lines to be used in the future. If
we do the math, this allows for a maximum of 1040 interrupt lines. That
many interrupt lines being available is not practical. We lower the reserved
chunk to 256 bytes, giving us a more reasonable maximum of 80.

Signed-off-by: Ian Moffett <ian@osmora.org>